Move Github client outside protocol directory
