+++ src/main/java/org/radarbase/appserver/controller/GithubEndpoint.java
@@ -22,7 +22,7 @@
 package org.radarbase.appserver.controller;
 
 import org.radarbase.appserver.config.AuthConfig;
-import org.radarbase.appserver.service.protocol.GithubClient;
+import org.radarbase.appserver.service.GithubClient;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -30,7 +30,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import radar.spring.auth.common.Authorized;
-import radar.spring.auth.common.PermissionOn;
 
 import java.io.IOException;
 import java.net.MalformedURLException;
